backend/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
opencv-python==4.8.0
retina-face==0.0.14
insightface==0.7.3
clip @ git+https://github.com/openai/CLIP.git
weaviate-client==3.25.0
python-dotenv==1.0.0
spacy==3.6.1
flask==2.3.2